weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "QWeather MCP Server - 气象数据查询服务"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.27.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
weather-mcp = "weather.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.hatchling]
package-dir = "src"
packages = ["weather"]
